Module datom::builtin_idents

IDs for the built-in attributes and other idents

Constants

CARDINALITY

Whether an attribute associates one value or multiple values with an entity

CARDINALITY_MANY

A value for the CARDINALITY attribute

CARDINALITY_ONE

A value for the CARDINALITY attribute

DOC

A documentation string for an entity

ID

An entity’s ID. This is a virtual attribute, and isn’t actually stored.

IDENT

An entity’s alias, usually used for the attribute schema.

IS_COMPONENT

Whether the entity referred to in this type_ref attribute is a sub-component. When you retract an entity, all sub-components will also be retracted.

TYPE_BOOLEAN

A value for the VALUE_TYPE attribute

TYPE_DECIMAL

A value for the VALUE_TYPE attribute

TYPE_ID

A value for the VALUE_TYPE attribute

TYPE_INTEGER

A value for the VALUE_TYPE attribute

TYPE_REF

A value for the VALUE_TYPE attribute

TYPE_STRING

A value for the VALUE_TYPE attribute

UNIQUE

Whether there can only be one entity per value for this attribute

VALUE_TYPE

The intended type for an attribute. Note that this is not (currently) strictly checked.

Statics

BUILTIN_ENTITIES

The data behind all built-in entities

BUILTIN_ENTITIES_BY_IDENT

The data behind all built-in entities, by ident

Type Definitions

BuiltinEntities

The type of the data behind all built-in entities

BuiltinEntitiesByIdent

The type of the data behind all built-in entities, by ident

BuiltinEntity

The data behind a built-in entity