Enum datom::EntityResult

pub enum EntityResult<'connection, S: Storage> {
    NotFound,
    Value(Value),
    Ref(Entity<'connection, S>),
    Repeated(Vec<Self>),
}

The result of getting an attribute on an entity

Variants

NotFound

A value for that attribute wasn’t found on this entity

Value(Value)

A value for that attribute was found on this entity

Ref(Entity<'connection, S>)

A value for that attribute was found on this entity, and it refers to another entity.

Repeated(Vec<Self>)

Possibly multiple values for that attribute were found on this entity

Implementations

impl<'connection, S: Storage> EntityResult<'connection, S>

pub fn is_ref_to(&self, id: &ID) -> bool

Check if the result is a reference to a specific ID