Module datom::serial

Serialization/deserialization functions

Functions

avet_attribute_value_range

Create a range encompassing every possible datom for a given attribute and value in the AVET index

deserialize

Deserialize a datom from a given index

deserialize_aevt

Deserialize a datom in attribute-entity-value-t order

deserialize_avet

Deserialize a datom in attribute-value-entity-t order

deserialize_eavt

Deserialize a datom in entity-attribute-value-t order

deserialize_tr

Deserialize a TransactionRecord

deserialize_unknown

Deserialize a datom from any index

deserialize_vaet

Deserialize a datom in value-attribute-entity-t order

eavt_entity_attribute_range

Create a range encompassing every possible datom for a given entity and attribute in the EAVT index

eavt_entity_range

Create a range encompassing every possible datom for a given entity in the EAVT index

index_range

Create a range encompassing an entire index

range_slice

Convert a range of arrays to a range of slices

serialize

Serialize a datom for a given index

serialize_aevt

Serialize a datom in attribute-entity-value-t order

serialize_avet

Serialize a datom in attribute-value-entity-t order

serialize_eavt

Serialize a datom in entity-attribute-value-t order

serialize_tr

Serialize a TransactionRecord

serialize_vaet

Serialize a datom in value-attribute-entity-t order

tr_range

Create a range encompassing every transaction result

vaet_value_attribute_range

Create a range encompassing every possible datom for a given value and attribute in the VAET index

vec_range_slice

Convert a range of Vecs to a range of slices