Crate datom

An open-source database inspired by Datomic.

// SPDX-FileCopyrightText: 2022 Lutris, Inc
// SPDX-License-Identifier: BlueOak-1.0.0 OR BSD-2-Clause-Patent
// SPDX-FileContributor: Piper McCorkle <piper@lutris.engineering>

Currently built on top of sled, a modern embedded database, and is only suitable for embedded usage. Future multi-peer support is planned.

If you aren’t already familiar with Datomic, it might help to look over Datomic’s excellent documentation. To the user, datom-rs functions extremely similarly to Datomic.

use datom::{backends::SledStorage, Connection, EntityResult, Transaction, Value, ID};

// Use the sled storage backend to create a temporary database
let storage = SledStorage::connect_temp()?;

// Create a connection from that backend
let conn = Connection::new(storage);

// Create an ID to use for the username attribute
let username = ID::new();
// Create an ID to use for the user's entity
let user = ID::new();

// Create a transaction setting the username attribute on the user
// entity to "pmc"
let mut tx = Transaction::new();
tx.add(user.into(), username.into(), "pmc".into());
// Execute the transaction using the connection
conn.transact(tx)?;

// Get a view of the database in the current point in time
let db = conn.db()?;
// Get the value of the username attribute on the user entity
if let EntityResult::Value(Value::String(u)) = db.entity(user.into())?.get(username.into())? {
    println!("The user's username is {}.", u);
}

Modules

backends

Storage backends

builtin_idents

IDs for the built-in attributes and other idents

serial

Serialization/deserialization functions

storage

API for storage backends

Structs

AttributeIterator

An iterator over attributes in a sled-backed database

AttributeSchema

An imperative way to generate an attribute’s schema

Connection

A persistent connection to a database

Database

A view of a database at a specific point in time

Datom

A datom, or a single fact at a single point in time. Short for data atom.

DatomIterator

An iterator over Datoms

Entity

An entity in a database

ID

An entity ID

Transaction

A set of facts which can be transacted into a database connection

TransactionRecord

The record of a past transaction

TransactionResult

The result of running a Transaction on a Connection

Enums

AttributeType

The type of an attribute’s values

ConnectionError

Network/disk errors

DatomType

Whether a datom is showing an addition or a retraction

EID

An un-resolved entity ID, which can be used to resolve entities by ident or unique attribute

EntityResult

The result of getting an attribute on an entity

Fact

A fact which hasn’t yet been converted to a Datom (or set of Datoms)

Index

The four indices used in the underlying data store. The names refer to the serialization order, and by extension the sort order.

QueryError

Errors during a Database query

StorageError

An error in the underlying storage backend

TransactionError

Errors during a Transaction

Value

An attribute value.

Traits

Transactable

A type which can be appended to a transaction

Functions

new_dynamic_connection

Create a new connection which uses a dynamically dispatched storage backend

parse_edn

Parse an EDN string to an Edn object

version

Get the version of this datom build

Type Aliases

DynamicConnection

A connection which uses a dynamically dispatched storage backend